requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
click==8.0.4
colorama==0.4.4
fastapi==0.74.1
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
protobuf==3.19.4
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.17.5
watchgod==0.7
websockets==10.2
